#283 refractor fact testing
