+++ src/bud/test_bud_settle/test_z_agenda.py
@@ -2,11 +2,8 @@ from datetime import datetime
 from src._road.road import RoadUnit
 from src.bud.bud import budunit_shop, get_from_json as budunit_get_from_json
 from src.bud.idea import IdeaUnit, ideaunit_shop
-from src.bud.reason_idea import reasonunit_shop
-from src.bud.group import awardlink_shop
 from src.bud.reason_doer import doerunit_shop
 from src.bud.examples.example_buds import (
-    get_budunit_1Task_1CE0MinutesReason_1Fact,
     get_budunit_with_4_levels,
     get_budunit_with_4_levels_and_2reasons,
     get_budunit_with7amCleanTableReason,
