+++ src/bud/test_bud_time/test_z_time_agenda.py
@@ -82,228 +82,229 @@ def test_get_time_min_from_year_ReturnsObj():
     assert _get_year_min_length(2004) == 527040
 
 
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_day_idea_Scenario0():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    day_road = sue_budunit.make_road(jaja_road, day_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # day_idea = sue_budunit.get_idea_obj(day_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_day_idea_Scenario0():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     day_road = sue_budunit.make_road(jaja_road, day_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # day_idea = sue_budunit.get_idea_obj(day_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
 
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
 
-    # THEN
-    assert sue_agenda.get(clean_road)
-
-
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_day_idea_Scenario1():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    day_road = sue_budunit.make_road(jaja_road, day_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # day_idea = sue_budunit.get_idea_obj(day_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
-    sue_budunit.edit_idea_attr(
-        clean_road,
-        reason_base=day_road,
-        reason_premise=day_road,
-        reason_premise_open=0,
-        reason_premise_nigh=1,
-        reason_premise_divisor=1,
-    )
-    sue_budunit.set_fact(day_road, day_road, 0, 1)
-
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
-    print(f"{sue_agenda=}")
-
-    # THEN
-    assert sue_agenda.get(clean_road)
+#     # THEN
+#     assert sue_agenda.get(clean_road)
 
 
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_day_idea_Scenario2():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    day_road = sue_budunit.make_road(jaja_road, day_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # day_idea = sue_budunit.get_idea_obj(day_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
-    sue_budunit.edit_idea_attr(
-        clean_road,
-        reason_base=day_road,
-        reason_premise=day_road,
-        reason_premise_open=360,
-        reason_premise_nigh=420,
-        reason_premise_divisor=1440,
-    )
-    sue_budunit.set_fact(day_road, day_road, 14400300, 14400480)
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_day_idea_Scenario1():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     day_road = sue_budunit.make_road(jaja_road, day_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # day_idea = sue_budunit.get_idea_obj(day_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+#     sue_budunit.edit_idea_attr(
+#         clean_road,
+#         reason_base=day_road,
+#         reason_premise=day_road,
+#         reason_premise_open=0,
+#         reason_premise_nigh=1,
+#         reason_premise_divisor=1,
+#     )
+#     sue_budunit.set_fact(jaja_road, jaja_road, 0, 1440)
 
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
+#     print(f"{sue_agenda=}")
 
-    # THEN
-    assert sue_agenda.get(clean_road) != None
-
-
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_days_idea_Scenario0():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    days_road = sue_budunit.make_road(jaja_road, days_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # days_idea = sue_budunit.get_idea_obj(days_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
-    sue_budunit.edit_idea_attr(
-        clean_road,
-        reason_base=days_road,
-        reason_premise=days_road,
-        reason_premise_open=4,
-        reason_premise_nigh=5,
-        reason_premise_divisor=7,
-    )
-    sue_budunit.set_fact(days_road, days_road, 11, 12)
-
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
-
-    # THEN
-    assert sue_agenda.get(clean_road)
+#     # THEN
+#     assert sue_agenda.get(clean_road)
 
 
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_week_idea_Scenario0():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    week_road = sue_budunit.make_road(jaja_road, week_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # week_idea = sue_budunit.get_idea_obj(week_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
-    sue_budunit.edit_idea_attr(
-        clean_road,
-        reason_base=week_road,
-        reason_premise=week_road,
-        reason_premise_open=0,
-        reason_premise_nigh=1440,
-        reason_premise_divisor=10080,
-    )
-    sue_budunit.set_fact(week_road, week_road, 0, 1440)
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_day_idea_Scenario2():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     day_road = sue_budunit.make_road(jaja_road, day_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # day_idea = sue_budunit.get_idea_obj(day_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+#     sue_budunit.edit_idea_attr(
+#         clean_road,
+#         reason_base=day_road,
+#         reason_premise=day_road,
+#         reason_premise_open=360,
+#         reason_premise_nigh=420,
+#         reason_premise_divisor=1440,
+#     )
+#     sue_budunit.set_fact(jaja_road, jaja_road, 14400300, 14400480)
 
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
-    print(f"{sue_agenda=}")
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
 
-    # THEN
-    assert sue_agenda.get(clean_road)
-
-
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_week_idea_Scenario1():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    week_road = sue_budunit.make_road(jaja_road, week_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # week_idea = sue_budunit.get_idea_obj(week_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
-    sue_budunit.edit_idea_attr(
-        clean_road,
-        reason_base=week_road,
-        reason_premise=week_road,
-        reason_premise_open=2880,
-        reason_premise_nigh=4220,
-        reason_premise_divisor=10080,
-    )
-    sue_budunit.set_fact(week_road, week_road, 100802880, 100804220)
-
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
-
-    # THEN
-    assert sue_agenda.get(clean_road) != None
+#     # THEN
+#     assert sue_agenda.get(clean_road) != None
 
 
-def test_BudUnit_get_agenda_dict_ReturnsDictWith_weeks_idea_Scenario0():
-    # ESTABLISH
-    sue_budunit = budunit_shop("Sue")
-    time_road = sue_budunit.make_l1_road(time_str())
-    jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
-    weeks_road = sue_budunit.make_road(jaja_road, weeks_str())
-    sue_budunit = add_time_hreg_ideaunit(sue_budunit)
-    # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
-    # weeks_idea = sue_budunit.get_idea_obj(weeks_road)
-    sue_budunit.tree_range_traverse_calc()
-    casa_text = "casa"
-    casa_road = sue_budunit.make_l1_road(casa_text)
-    clean_text = "clean"
-    clean_road = sue_budunit.make_road(casa_road, clean_text)
-    sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
-    sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
-    sue_budunit.edit_idea_attr(
-        clean_road,
-        reason_base=weeks_road,
-        reason_premise=weeks_road,
-        reason_premise_open=4,
-        reason_premise_nigh=5,
-        reason_premise_divisor=7,
-    )
-    sue_budunit.set_fact(weeks_road, weeks_road, 11, 12)
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_days_idea_Scenario0():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     days_road = sue_budunit.make_road(jaja_road, days_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # days_idea = sue_budunit.get_idea_obj(days_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+#     sue_budunit.edit_idea_attr(
+#         clean_road,
+#         reason_base=days_road,
+#         reason_premise=days_road,
+#         reason_premise_open=4,
+#         reason_premise_nigh=5,
+#         reason_premise_divisor=7,
+#     )
+#     sue_budunit.set_fact(jaja_road, jaja_road, 11, 12)
 
-    # WHEN
-    sue_agenda = sue_budunit.get_agenda_dict()
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
 
-    # THEN
-    assert sue_agenda.get(clean_road)
+#     # THEN
+#     assert sue_agenda.get(clean_road)
+
+
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_week_idea_Scenario0():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     week_road = sue_budunit.make_road(jaja_road, week_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # week_idea = sue_budunit.get_idea_obj(week_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+#     sue_budunit.edit_idea_attr(
+#         clean_road,
+#         reason_base=week_road,
+#         reason_premise=week_road,
+#         reason_premise_open=0,
+#         reason_premise_nigh=1440,
+#         reason_premise_divisor=10080,
+#     )
+#     sue_budunit.set_fact(jaja_road, jaja_road, 0, 1440)
+
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
+#     print(f"{sue_agenda=}")
+
+#     # THEN
+#     assert sue_agenda.get(clean_road)
+
+
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_week_idea_Scenario1():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     week_road = sue_budunit.make_road(jaja_road, week_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # week_idea = sue_budunit.get_idea_obj(week_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+#     sue_budunit.edit_idea_attr(
+#         clean_road,
+#         reason_base=week_road,
+#         reason_premise=week_road,
+#         reason_premise_open=2880,
+#         reason_premise_nigh=4220,
+#         reason_premise_divisor=10080,
+#     )
+#     sue_budunit.set_fact(jaja_road, jaja_road, 100802880, 100804220)
+
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
+
+#     # THEN
+#     assert sue_agenda.get(clean_road) != None
+
+
+# def test_BudUnit_get_agenda_dict_ReturnsDictWith_weeks_idea_Scenario0():
+#     # ESTABLISH
+#     sue_budunit = budunit_shop("Sue")
+#     time_road = sue_budunit.make_l1_road(time_str())
+#     jaja_road = sue_budunit.make_road(time_road, get_jajatime_text())
+#     weeks_road = sue_budunit.make_road(jaja_road, weeks_str())
+#     sue_budunit = add_time_hreg_ideaunit(sue_budunit)
+#     # jaja_idea = sue_budunit.get_idea_obj(jaja_road)
+#     # weeks_idea = sue_budunit.get_idea_obj(weeks_road)
+#     sue_budunit.tree_range_traverse_calc()
+#     casa_text = "casa"
+#     casa_road = sue_budunit.make_l1_road(casa_text)
+#     clean_text = "clean"
+#     clean_road = sue_budunit.make_road(casa_road, clean_text)
+#     sue_budunit.set_l1_idea(ideaunit_shop(casa_text))
+#     sue_budunit.set_idea(ideaunit_shop(clean_text, pledge=True), casa_road)
+#     sue_budunit.edit_idea_attr(
+#         clean_road,
+#         reason_base=weeks_road,
+#         reason_premise=weeks_road,
+#         reason_premise_open=4,
+#         reason_premise_nigh=5,
+#         reason_premise_divisor=7,
+#     )
+#     sue_budunit.set_fact(weeks_road, weeks_road, 11, 12)
+
+#     # WHEN
+#     sue_agenda = sue_budunit.get_agenda_dict()
+
+#     # THEN
+#     assert sue_agenda.get(clean_road)
+#     assert 1 == 4
 
 
 # def test_BudUnit_get_agenda_dict_ReturnsDictWith_year_idea_Scenario0():
